rename profile component
